backend: Move ast.py to common/astutils.py
